Refactors signal forwarding and makes Windows-friendly.
